refactor(text): consolidate string truncation on truncate_chars
Route the remaining hand-rolled truncators onto the shared
character-based helper:

- promote truncate_chars from app/util.rs to the crate-level text
  module so modules outside app (search_overlay) can share it
- ssh upload file names: drop truncate_name (ASCII "...") for
  truncate_chars
- search overlay detail rows: truncate_to_width keeps the
  width-to-chars estimate but delegates the cut to truncate_chars
  (sub-character budgets keep the legacy ellipsis placeholder)
- horizon-core: new util::truncate_chars (mirrors the UI helper;
  core cannot depend on the UI) and agent session labels now use
  the single-char ellipsis instead of ASCII "..."

Part of #271 (item 2). Stacked on #586, which moves the
panel_chrome call site to the new home.
