+++ crates/horizon-core/src/util.rs
@@ -1,0 +1,61 @@
+//! Small shared helpers with no more specific module.
+
+use std::borrow::Cow;
+
+/// Caps `value` at `max_chars` characters, ending a truncated result with an
+/// ellipsis that counts toward the budget. Borrows when nothing is cut.
+///
+/// The cut is character-based: byte-indexed slicing would panic whenever a
+/// multi-byte character straddles the cut point. Mirrors the UI's
+/// `crate::text::truncate_chars`; horizon-core cannot depend on the UI, so
+/// keep the two implementations in sync.
+pub(crate) fn truncate_chars(value: &str, max_chars: usize) -> Cow<'_, str> {
+    if max_chars == 0 {
+        return Cow::Borrowed(if value.is_empty() { value } else { "" });
+    }
+
+    let mut ellipsis_at = 0;
+    for (char_index, (byte_index, _)) in value.char_indices().enumerate() {
+        if char_index + 1 == max_chars {
+            ellipsis_at = byte_index;
+        } else if char_index + 1 > max_chars {
+            let mut truncated = String::with_capacity(ellipsis_at + '…'.len_utf8());
+            truncated.push_str(&value[..ellipsis_at]);
+            truncated.push('…');
+            return Cow::Owned(truncated);
+        }
+    }
+
+    Cow::Borrowed(value)
+}
+
+#[cfg(test)]
+mod tests {
+    use super::truncate_chars;
+
+    #[test]
+    fn truncate_chars_keeps_short_values_untouched() {
+        assert_eq!(truncate_chars("build", 8), "build");
+        assert_eq!(truncate_chars("exactly8", 8), "exactly8");
+    }
+
+    #[test]
+    fn truncate_chars_caps_at_budget_including_ellipsis() {
+        let truncated = truncate_chars("docker compose up", 8);
+
+        assert_eq!(truncated, "docker …");
+        assert_eq!(truncated.chars().count(), 8);
+    }
+
+    #[test]
+    fn truncate_chars_honors_tiny_budgets() {
+        assert_eq!(truncate_chars("abc", 0), "");
+        assert_eq!(truncate_chars("abc", 1), "…");
+        assert_eq!(truncate_chars("", 0), "");
+    }
+
+    #[test]
+    fn truncate_chars_counts_characters_not_bytes() {
+        assert_eq!(truncate_chars("blåbærsyltetøy", 6), "blåbæ…");
+    }
+}
