+++ crates/horizon-ui/src/app/util.rs
@@ -96,29 +96,7 @@ pub(super) fn editor_panel_size_for_file(path: &std::path::Path) -> [f32; 2] {
 }
 
 pub(super) fn truncate_session_label(label: &str) -> String {
-    truncate_chars(label, 40).into_owned()
-}
-
-/// Caps `label` at `max_chars` characters, ending a truncated result with an
-/// ellipsis that counts toward the budget. Borrows when nothing is cut.
-pub(super) fn truncate_chars(label: &str, max_chars: usize) -> std::borrow::Cow<'_, str> {
-    if max_chars == 0 {
-        return std::borrow::Cow::Borrowed(if label.is_empty() { label } else { "" });
-    }
-
-    let mut ellipsis_at = 0;
-    for (char_index, (byte_index, _)) in label.char_indices().enumerate() {
-        if char_index + 1 == max_chars {
-            ellipsis_at = byte_index;
-        } else if char_index + 1 > max_chars {
-            let mut truncated = String::with_capacity(ellipsis_at + '…'.len_utf8());
-            truncated.push_str(&label[..ellipsis_at]);
-            truncated.push('…');
-            return std::borrow::Cow::Owned(truncated);
-        }
-    }
-
-    std::borrow::Cow::Borrowed(label)
+    crate::text::truncate_chars(label, 40).into_owned()
 }
 
 pub(super) fn paint_canvas_glow(ui: &mut egui::Ui) {
@@ -238,34 +216,8 @@ pub(super) fn atomic_write(path: &std::path::Path, content: &str) -> std::io::Re
 
 #[cfg(test)]
 mod tests {
-    use super::{OverlayExclusion, clamp_panel_size, format_grid_position, primary_shortcut_label, truncate_chars};
+    use super::{OverlayExclusion, clamp_panel_size, format_grid_position, primary_shortcut_label};
     use egui::{Pos2, Rect, Vec2};
-
-    #[test]
-    fn truncate_chars_keeps_short_labels_untouched() {
-        assert_eq!(truncate_chars("build", 8), "build");
-        assert_eq!(truncate_chars("exactly8", 8), "exactly8");
-    }
-
-    #[test]
-    fn truncate_chars_caps_at_budget_including_ellipsis() {
-        let truncated = truncate_chars("docker compose up", 8);
-
-        assert_eq!(truncated, "docker …");
-        assert_eq!(truncated.chars().count(), 8);
-    }
-
-    #[test]
-    fn truncate_chars_honors_tiny_budgets() {
-        assert_eq!(truncate_chars("abc", 0), "");
-        assert_eq!(truncate_chars("abc", 1), "…");
-        assert_eq!(truncate_chars("", 0), "");
-    }
-
-    #[test]
-    fn truncate_chars_counts_characters_not_bytes() {
-        assert_eq!(truncate_chars("blåbærsyltetøy", 6), "blåbæ…");
-    }
 
     fn default_panel_canvas_pos(index: usize) -> Pos2 {
         const PANEL_COLUMN_SPACING: f32 = 540.0;
